+++ src/Tests/AppInsights.fs
@@ -3,6 +3,8 @@ module AppInsights
 open Expecto
 open Farmer
 open Farmer.Builders.AppInsights
+open Farmer.Builders.LogAnalytics
+open Newtonsoft.Json.Linq
 
 let tests = testList "AppInsights" [
     test "Creates keys on an AI instance correctly" {
@@ -11,8 +13,31 @@ let tests = testList "AppInsights" [
         Expect.equal ai.InstrumentationKey.Value ("reference(resourceId('Microsoft.Insights/components', 'foo'), '2014-04-01').InstrumentationKey") "Incorrect Value"
     }
 
+    test "Creates with classic version by default" {
+        let deployment = arm { add_resource (appInsights { name "foo" }) }
+        let json = deployment.Template |> Writer.toJson |> JObject.Parse
+        let version = json.SelectToken("resources[?(@.name=='foo')].apiVersion").ToString()
+        Expect.equal version "2014-04-01" "Incorrect API version"
+    }
+
     test "Create generated keys correctly" {
         let generatedKey = AppInsights.getInstrumentationKey(ResourceId.create(Arm.Insights.components, ResourceName "foo", "group"))
         Expect.equal generatedKey.Value "reference(resourceId('group', 'Microsoft.Insights/components', 'foo'), '2014-04-01').InstrumentationKey" "Incorrect generated key"
     }
+
+    test "Creates LA-enabled workspace" {
+        let workspace = logAnalytics { name "la" }
+        let ai = appInsights { name "ai"; log_analytics_workspace workspace }
+        let deployment = arm {
+            add_resources [ workspace; ai ]
+        }
+
+        let json = deployment.Template |> Writer.toJson |> JObject.Parse
+        let select query = json.SelectToken(query).ToString()
+
+        Expect.equal (select "resources[?(@.name=='ai')].properties.WorkspaceResourceId") "[resourceId('Microsoft.OperationalInsights/workspaces', 'la')]" "Incorrect workspace id"
+        Expect.equal (select "resources[?(@.name=='ai')].apiVersion") "2020-02-02" "Incorrect API version"
+        Expect.equal ai.InstrumentationKey.Value ("reference(resourceId('Microsoft.Insights/components', 'ai'), '2020-02-02').InstrumentationKey") "Incorrect Instrumentation Key reference"
+        Expect.sequenceEqual (json.SelectToken("resources[?(@.name=='ai')].dependsOn").Children() |> Seq.map string |> Seq.toArray) [ "[resourceId('Microsoft.OperationalInsights/workspaces', 'la')]" ] "Incorrect dependencies"
+   }
 ]