+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 Release Notes
 =============
 ## 1.6.30
+* WebApps/Functions: Specify connection string types
 * WebApps/Functions: Allow adding IP restriction string with CIDR
 * Application Insights: Support for Workspace-enabled instances.