+++ RELEASE_NOTES.md
@@ -2,11 +2,12 @@ Release Notes
 =============
 
 ## vNext
-* WebApps/Functions: Specify connection string types
 * VMs: Priority and Spot Instance Settings
 
 ## 1.6.30
+* WebApps/Functions: Specify connection string types
 * WebApps/Functions: Allow adding IP restriction string with CIDR
+* Application Insights: Support for Workspace-enabled instances.
 
 ## 1.6.29
 * CLI: include `--only-show-error` option when executing Azure CLI commands.