+++ RELEASE_NOTES.md
@@ -1,10 +1,15 @@
 Release Notes
 =============
 
-## 1.6.30
-* WebApps/Functions: Allow adding IP restriction string with CIDR
-* WebApps/Functions: Specify connection string types
+## 1.6.31
 * WebApps/Functions: Add support for vnet integration
+
+
+## 1.6.30
+* WebApps/Functions: Specify connection string types
+* WebApps/Functions: Allow adding IP restriction string with CIDR
+* Application Insights: Support for Workspace-enabled instances.
+* VMs: Priority and Spot Instance Settings
 
 ## 1.6.29
 * CLI: include `--only-show-error` option when executing Azure CLI commands.